Model ve Controller'larda geliştirme yapıldı. 06.05.2022

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -142,7 +142,7 @@ class CategoryController extends Controller
     public function destroy(category $categories, $id)
     {
         $data=category::find($id);
-        if($data->image)
+        if($data->image && Storage::disk('public')->exists($data->image))
             Storage::delete($data->image);
         $data->delete();
         return redirect('/admin/category');
